+++ test/hierholzer_trail_test.cpp
@@ -7,14 +7,14 @@
 #include "../src/trail/trailstructure.h"
 
 TEST(EulerTrailTest, hierholzerTrail) {
-    uint32_t order = 6;
-    auto **adj_mtrx = reinterpret_cast<uint32_t **>(malloc(sizeof(uint32_t) * order * order));
-    adj_mtrx[0] = new uint32_t[order]{0, 1, 0, 1, 1, 1};
-    adj_mtrx[1] = new uint32_t[order]{1, 0, 1, 0, 0, 0};
-    adj_mtrx[2] = new uint32_t[order]{0, 1, 0, 1, 0, 0};
-    adj_mtrx[3] = new uint32_t[order]{1, 0, 1, 0, 0, 0};
-    adj_mtrx[4] = new uint32_t[order]{1, 0, 0, 0, 0, 0};
-    adj_mtrx[5] = new uint32_t[order]{1, 0, 0, 0, 0, 0};
+    uint64_t order = 6;
+    auto **adj_mtrx = reinterpret_cast<uint64_t **>(malloc(sizeof(uint64_t) * order * order));
+    adj_mtrx[0] = new uint64_t[order]{0, 1, 0, 1, 1, 1};
+    adj_mtrx[1] = new uint64_t[order]{1, 0, 1, 0, 0, 0};
+    adj_mtrx[2] = new uint64_t[order]{0, 1, 0, 1, 0, 0};
+    adj_mtrx[3] = new uint64_t[order]{1, 0, 1, 0, 0, 0};
+    adj_mtrx[4] = new uint64_t[order]{1, 0, 0, 0, 0, 0};
+    adj_mtrx[5] = new uint64_t[order]{1, 0, 0, 0, 0, 0};
 
     std::shared_ptr<Sealib::UndirectedGraph>
         graph_ptr(
